Changing strings to tuples
In the NETS and DATASETS dictionary, we have only a single element inside the tuple. Hence, it will be treated as strings, and not tuples. 

Hence, 
`tfmodel = os.path.join('../output',demonet,DATASETS[dataset][0], 'default', NETS[demonet][0]) `
in line 119 will compute :
`../output/res101/v/default/r.meta`

Hence the IOError in 124 fires, even if the data is present in the path and arguments are passed correctly. 

FIX would be to explicitly say that it is a tuple.

Ref: http://stackoverflow.com/questions/12876177/why-do-tuples-with-only-one-element-get-converted-to-strings